project with docker files and mqtt protocol
