+++ .vscode/launch.json
@@ -1,60 +1,65 @@
+// AUTOMATICALLY GENERATED FILE. PLEASE DO NOT MODIFY IT MANUALLY
+//
+// PIO Unified Debugger
+//
+// Documentation: https://docs.platformio.org/page/plus/debugging.html
+// Configuration: https://docs.platformio.org/page/projectconf/section_env_debug.html
+
 {
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "name": "C/C++ Runner: Debug Session",
-      "type": "cppdbg",
-      "request": "launch",
-      "args": [
-        ""
-      ],
-      "stopAtEntry": false,
-      "cwd": "/home/ruben/repos/ru/iot-mqtt-client",
-      "environment": [],
-      "program": "/home/ruben/repos/ru/iot-mqtt-client/build/Debug/outDebug",
-      "internalConsoleOptions": "openOnSessionStart",
-      "MIMode": "gdb",
-      "miDebuggerPath": "/usr/bin/gdb",
-      "externalConsole": false,
-      "setupCommands": [
+    "version": "0.2.0",
+    "configurations": [
         {
-          "description": "Enable pretty-printing for gdb",
-          "text": "-enable-pretty-printing",
-          "ignoreFailures": true
+            "name": "C/C++ Runner: Debug Session",
+            "type": "cppdbg",
+            "request": "launch",
+            "args": [],
+            "stopAtEntry": false,
+            "cwd": "/home/ruben/repos/ru/iot-mqtt-client",
+            "environment": [],
+            "program": "/home/ruben/repos/ru/iot-mqtt-client/build/Debug/outDebug",
+            "internalConsoleOptions": "openOnSessionStart",
+            "MIMode": "gdb",
+            "miDebuggerPath": "/usr/bin/gdb",
+            "externalConsole": false,
+            "setupCommands": [
+                {
+                    "description": "Enable pretty-printing for gdb",
+                    "text": "-enable-pretty-printing",
+                    "ignoreFailures": true
+                }
+            ]
+        },
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug",
+            "executable": "/home/ruben/repos/ru/iot-mqtt-client/.pio/build/esp32dev/firmware.elf",
+            "projectEnvName": "esp32dev",
+            "toolchainBinDir": "/home/ruben/.platformio/packages/toolchain-xtensa32/bin",
+            "internalConsoleOptions": "openOnSessionStart",
+            "preLaunchTask": {
+                "type": "PlatformIO",
+                "task": "Pre-Debug"
+            }
+        },
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug (skip Pre-Debug)",
+            "executable": "/home/ruben/repos/ru/iot-mqtt-client/.pio/build/esp32dev/firmware.elf",
+            "projectEnvName": "esp32dev",
+            "toolchainBinDir": "/home/ruben/.platformio/packages/toolchain-xtensa32/bin",
+            "internalConsoleOptions": "openOnSessionStart"
+        },
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug (without uploading)",
+            "executable": "/home/ruben/repos/ru/iot-mqtt-client/.pio/build/esp32dev/firmware.elf",
+            "projectEnvName": "esp32dev",
+            "toolchainBinDir": "/home/ruben/.platformio/packages/toolchain-xtensa32/bin",
+            "internalConsoleOptions": "openOnSessionStart",
+            "loadMode": "manual"
         }
-      ]
-    },
-    {
-      "type": "platformio-debug",
-      "request": "launch",
-      "name": "PIO Debug",
-      "executable": "/home/ruben/repos/ru/iot-mqtt-client/.pio/build/esp32dev/firmware.elf",
-      "projectEnvName": "esp32dev",
-      "toolchainBinDir": "/home/ruben/.platformio/packages/toolchain-xtensa32/bin",
-      "internalConsoleOptions": "openOnSessionStart",
-      "preLaunchTask": {
-        "type": "PlatformIO",
-        "task": "Pre-Debug"
-      }
-    },
-    {
-      "type": "platformio-debug",
-      "request": "launch",
-      "name": "PIO Debug (skip Pre-Debug)",
-      "executable": "/home/ruben/repos/ru/iot-mqtt-client/.pio/build/esp32dev/firmware.elf",
-      "projectEnvName": "esp32dev",
-      "toolchainBinDir": "/home/ruben/.platformio/packages/toolchain-xtensa32/bin",
-      "internalConsoleOptions": "openOnSessionStart"
-    },
-    {
-      "type": "platformio-debug",
-      "request": "launch",
-      "name": "PIO Debug (without uploading)",
-      "executable": "/home/ruben/repos/ru/iot-mqtt-client/.pio/build/esp32dev/firmware.elf",
-      "projectEnvName": "esp32dev",
-      "toolchainBinDir": "/home/ruben/.platformio/packages/toolchain-xtensa32/bin",
-      "internalConsoleOptions": "openOnSessionStart",
-      "loadMode": "manual"
-    }
-  ]
+    ]
 }
