+++ docker/docker-compose.yml
@@ -1,0 +1,24 @@
+services:
+
+  db:
+    image: postgres
+    restart: always
+    ports:
+      - 5432:5432
+    environment:
+      POSTGRES_PASSWORD: 1234
+      POSTGRES_USER: postgres
+
+  adminer:
+    image: adminer
+    restart: always
+    ports:
+      - 8080:8080
+
+  broker:
+    image: emqx/emqx
+    restart: always
+    ports:
+      - 18083:18083
+      - 1883:1883
+      - 8083:8083
